Switch to map of public & internal instances
This simplifies the code a lot as we no longer rely on some unique key
in the internal instance (and also means it works for the production
build of react without the `_debugID`).

However, it does mean that the runtime needs a `Map()` that supports
objects as keys (polyfills may have performance considerations). This
should be ok on Node 4 and up.